Move zcash transaction_builder tests to forks/
Reduce diff to upstream
