+++ test/transaction_builder.js
@@ -14,7 +14,6 @@ var TransactionBuilder = require('../src/transaction_builder')
 var NETWORKS = require('../src/networks')
 
 var fixtures = require('./fixtures/transaction_builder')
-var fixturesZec = require('./fixtures/forks/zcash/transaction_builder')
 
 function construct (f, dontSign) {
   var network = NETWORKS[f.network]
@@ -134,59 +133,6 @@ describe('TransactionBuilder', function () {
         txAfter.outs.forEach(function (output, i) {
           assert.equal(bscript.toASM(output.script), f.outputs[i].script)
         })
-      })
-    })
-
-    fixturesZec.valid.forEach(function (testData) {
-      it('returns TransactionBuilder, with ' + testData.description, function () {
-        var network = NETWORKS['zcash']
-        var tx = Transaction.fromHex(testData.hex, network)
-        var txb = TransactionBuilder.fromTransaction(tx, network)
-
-        assert.equal(txb.tx.version, testData.version)
-        assert.equal(txb.tx.versionGroupId, testData.versionGroupId)
-        assert.equal(txb.tx.overwintered, testData.overwintered)
-        assert.equal(txb.tx.locktime, testData.locktime)
-        assert.equal(txb.tx.expiryHeight, testData.expiryHeight)
-        assert.equal(txb.tx.ins.length, testData.insLength)
-        assert.equal(txb.tx.outs.length, testData.outsLength)
-        assert.equal(txb.tx.joinsplits.length, testData.joinsplitsLength)
-        assert.equal(txb.tx.joinsplitPubkey.length, testData.joinsplitPubkeyLength)
-        assert.equal(txb.tx.joinsplitSig.length, testData.joinsplitSigLength)
-      })
-
-      it('throws if transaction builder network is incompatible for ' + testData.description, function () {
-        var errorMessage = 'This transaction is incompatible with the transaction builder'
-        var tx = Transaction.fromHex(testData.hex, NETWORKS.zcash)
-
-        // Zcash transaction but different network
-        assert.throws(function () {
-          TransactionBuilder.fromTransaction(tx)
-        }, new RegExp(errorMessage))
-
-        assert.throws(function () {
-          TransactionBuilder.fromTransaction(tx, NETWORKS.bitcoincash)
-        }, new RegExp(errorMessage))
-
-        assert.throws(function () {
-          TransactionBuilder.fromTransaction(tx, NETWORKS.bitcoinsv)
-        }, new RegExp(errorMessage))
-
-        assert.throws(function () {
-          TransactionBuilder.fromTransaction(tx, NETWORKS.bitcoingold)
-        }, new RegExp(errorMessage))
-
-        assert.throws(function () {
-          TransactionBuilder.fromTransaction(tx, NETWORKS.bitcoin)
-        }, new RegExp(errorMessage))
-
-        assert.throws(function () {
-          TransactionBuilder.fromTransaction(tx, NETWORKS.litecoin)
-        }, new RegExp(errorMessage))
-
-        assert.throws(function () {
-          TransactionBuilder.fromTransaction(tx, NETWORKS.testnet)
-        }, new RegExp(errorMessage))
       })
     })
